Char & Str now accepted parameters for functions. Strs use 2 registers for their information, one for their bytes, the other for the length. This is mostly to cut down on not needing to rescan the length of a string every operation. O(1) time versus O(n) time. Removed type_checker influence from ASMGenerator's functions. Will continue to remove the class and ensure nothing breaks with next update. Makefile update, now uses 2 different sets of compiler flags depending on if building for dev build or for live build.

+++ src/ErrAndRep/ErrorHandler.h
@@ -56,7 +56,7 @@ private:
    void report_one(const std::string&, const std::string&, const Diagnostic&) const;
 
    std::unordered_map<CompPhase, std::vector<Diagnostic>> m_errors;
-   unsigned int err_count = 0;
+   long unsigned int err_count = 0;
 };
